// src/types.ts
export interface Job {
  id: string;
  title: string;
  company: string;
  logo: string;
  location: string;
  type: string;
  salary: number;
  currency: string;
  description: string;
  isBookMarked: boolean;
  experienceLevel: string;
  postedAt: string;
  rating?: number;
  applicants?: number;
  responsibilities?: string[];
}

export type Page = 'home' | 'details' | 'login' | 'signup' | 'applications' | 'companies' | 'contact';